fix(solace-message-client): disconnect from the broker gracefully
Gracefully means waiting for the 'DISCONNECT' confirmation event before disposing the session so that the broker can cleanup resources accordingly.
Also, when the session dies and cannot be recovered, we do not initiate a disconnect handshake with the broker, but instead immediately dispose the session.
In the spec, we removed "fakeAsync" zones to effectively wait for microtasks to complete.