feat(solace-message-client): allow consuming messages sent to a queue or topic endpoint
Added following API for consuming messages sent to a queue or topic endpoint: `SolaceMessageClient#consume$`

If passing a `string` literal, then it is used as the subscription topic to create a topic endpoint for, allowing messages to be reliably received even if the connection is unstable.
If passing a descriptor object, it is used as the config to connect to a queue or topic endpoint.

You can subscribe to a non-durable topic endpoint as following: `solaceMessageClient.consume$('topic');`

Which is equivalent to passing the following config:

 ```ts
 solaceMessageClient.consume$({
    topicEndpointSubscription: SolaceObjectFactory.createTopicDestination('topic'),
    queueDescriptor: {
      type: QueueType.TOPIC_ENDPOINT,
      durable: false,
    },
  });
 ```

To consume messages sent to a durable queue, pass the following config:

```ts
 solaceMessageClient.consume$({
   queueDescriptor: {
     type: QueueType.QUEUE,
     name: 'queue',
   },
 });
```