ci(solace-message-client): fix end-to-end test setup to start the application using `http-server`
In the GitHub Actions workflow, we build library and "TryMe" application as one of the first steps. When running the end-to-end tests, we start the application using `http-server` in order to speed up test execution. Since the "TryMe" application has a context-path set (required for deployment to GitHub Pages), we need to place the application in that context path as well, otherwise resources would not be found.