test: refactor tests as prerequisite to split tests into separate files
The large number of tests in a single file makes it difficult to maintain and add new tests. Initializing the session in 'beforeEach' blocks makes it difficult to get an overview and prevents different setups. As a first step, the fixtures were prepared for reuse and moved to separate files.